Fleshed out welcome and FAQ (#4)
* Removed getting started and related topics

Getting started and related topics like how to add a file/folder, downloading/uploading content must be in their respective file. Changed FAQs to actually be questions.

* Added getting started to welcome

This feels much more appropriate, this is where the docs begin and as such must introduce any developer directly to the basics.

* Deemed irrelevant, as there's no purpose

An introduction to what Koji is and dedicating an entire page on the docs for it violates the purpose of the documentation. Such introductions should be done in the homepage of the website. 
The welcome page subtely introduces what koji is.

* added a brief introduction to what koji is

* Removed what_is_koji

Removed the what_is_koji sections to reflect the updates made